backend/requirements.txt
flask
flask-cors
pymongo
python-dotenv
flask-jwt-extended
flask-socketio